requirements.txt
aiohappyeyeballs==2.3.7
aiohttp==3.10.4
aiosignal==1.3.1
altair==5.4.0
annotated-types==0.7.0
anyio==4.4.0
asttokens==2.4.1
async-timeout==4.0.3
attrs==24.2.0
blinker==1.8.2
cachetools==5.4.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.2.2
dataclasses-json==0.6.7
datasets==3.0.0
debugpy==1.8.5
decorator==5.1.1
dill==0.3.8
environs==9.5.0
eval_type_backport==0.2.0
exceptiongroup==1.2.2
executing==2.0.1
faiss-cpu==1.8.0.post1
fastapi==0.112.1
filelock==3.15.4
frozenlist==1.4.1
fsspec==2024.6.1
gitdb==4.0.11
GitPython==3.1.43
google-ai-generativelanguage==0.6.6
google-api-core==2.19.1
google-api-python-client==2.141.0
google-auth==2.33.0
google-auth-httplib2==0.2.0
google-generativeai==0.7.2
googleapis-common-protos==1.63.2
greenlet==3.0.3
grpcio==1.63.0
grpcio-status==1.62.3
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.2
huggingface-hub==0.24.5
idna==3.7
imgurpython==1.1.7
ipykernel==6.29.5
ipython==8.26.0
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.2
jupyter_core==5.7.2
langchain==0.2.14
langchain-community==0.2.12
langchain-core==0.2.33
langchain-huggingface==0.0.3
langchain-milvus==0.1.4
langchain-text-splitters==0.2.2
langsmith==0.1.99
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.3
matplotlib-inline==0.1.7
mdurl==0.1.2
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.16
mypy-extensions==1.0.0
narwhals==1.4.2
nest-asyncio==1.6.0
networkx==3.3
numpy==1.26.4
orjson==3.10.7
packaging==24.1
pandas==2.2.2
parso==0.8.4
pillow==10.4.0
platformdirs==4.2.2
prompt_toolkit==3.0.47
proto-plus==1.24.0
protobuf==4.25.4
psutil==6.0.0
pure_eval==0.2.3
pyarrow==17.0.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.8.2
pydantic_core==2.20.1
pydeck==0.9.1
Pygments==2.18.0
pymilvus==2.4.5
PyMuPDF==1.24.10
PyMuPDFb==1.24.10
pyparsing==3.1.2
pypdf==4.3.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
python-telegram-bot==21.5
pytz==2024.1
pywin32==306
PyYAML==6.0.2
pyzmq==26.1.0
referencing==0.35.1
regex==2024.7.24
requests==2.32.3
rich==13.7.1
rpds-py==0.20.0
rsa==4.9
safetensors==0.4.4
scikit-learn==1.5.1
scipy==1.14.0
sentence-transformers==3.0.1
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
SQLAlchemy==2.0.32
stack-data==0.6.3
starlette==0.38.2
streamlit==1.37.1
sympy==1.13.2
tabulate==0.9.0
telegram==0.0.1
tenacity==8.5.0
threadpoolctl==3.5.0
together==1.2.12
tokenizers==0.19.1
toml==0.10.2
torch==2.2.2+cpu
torchaudio==2.2.2+cpu
torchvision==0.17.2+cpu
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
transformers==4.44.0
typer==0.12.5
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
ujson==5.10.0
uritemplate==4.1.1
urllib3==2.2.2
uvicorn==0.30.6
watchdog==4.0.2
wcwidth==0.2.13
xxhash==3.5.0
yarl==1.9.4